requirements.txt

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
pyyaml==6.0.2
streamlit==1.37.0

reportlab==4.2.2
